requirements.txt
Django==3.2.5
django_filter==2.4.0
requests==2.23.0
pytz==2020.1
djangorestframework==3.12.4
